+++ src/js/entities-service/clinicians.js
@@ -1,21 +1,22 @@
-import Store from 'backbone.store';
-import BaseCollection from 'js/base/collection';
-import BaseModel from 'js/base/model';
+import BaseEntity from 'js/base/entity-service';
+import { _Model, Model, Collection } from './entities/clinicians';
 
-const TYPE = 'clinicians';
+const Entity = BaseEntity.extend({
+  Entity: { _Model, Model, Collection },
+  radioRequests: {
+    'clinicians:model': 'getModel',
+    'clinicians:collection': 'getCollection',
+    'fetch:clinician:model': 'fetchClinician',
+  },
+  fetchClinician(id) {
+    const include = [
+      'role',
+    ];
 
-const _Model = BaseModel.extend({
-  type: TYPE,
-  urlRoot: '/api/clinicians',
+    const data = { include };
+
+    return this.fetchModel(id, { data });
+  },
 });
 
-const Model = Store(_Model, TYPE);
-const Collection = BaseCollection.extend({
-  model: Model,
-});
-
-export {
-  _Model,
-  Model,
-  Collection,
-};
+export default new Entity();
