+++ src/js/entities-service/roles.js
@@ -1,21 +1,12 @@
-import Store from 'backbone.store';
-import BaseCollection from 'js/base/collection';
-import BaseModel from 'js/base/model';
+import BaseEntity from 'js/base/entity-service';
+import { _Model, Model, Collection } from './entities/roles';
 
-const TYPE = 'roles';
-
-const _Model = BaseModel.extend({
-  type: TYPE,
-  urlRoot: '/api/roles',
+const Entity = BaseEntity.extend({
+  Entity: { _Model, Model, Collection },
+  radioRequests: {
+    'roles:model': 'getModel',
+    'roles:collection': 'getCollection',
+  },
 });
 
-const Model = Store(_Model, TYPE);
-const Collection = BaseCollection.extend({
-  model: Model,
-});
-
-export {
-  _Model,
-  Model,
-  Collection,
-};
+export default new Entity();
